flushed out basic card wrapper to nail down design styles. this may be deleted as more robust future card components get built.
