+++ src/@primitives/CardWrapper/CardWrapper.stories.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { View } from 'react-native';
 import { storiesOf } from '@storybook/react-native';
 
-
 import Card from './';
 
 storiesOf('@primitives/Card', module)
@@ -11,11 +10,17 @@ storiesOf('@primitives/Card', module)
       justifyContent: 'center',
       alignItems: 'center',
       flex: 1,
+      backgroundColor: '#f7f7f7',
+    };
+
+    const cardDimensions = {
+      height: 400,
+      width: '92%',
     };
 
     return (
       <View style={centered}>
-        <Card />
+        <Card style={cardDimensions} />
       </View>
     );
   });
